// web/src/api/generated/model/workflowTaskDefSchemaInput.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.7.5
 */
import type { WorkflowTaskDefSchemaInputConfig } from './workflowTaskDefSchemaInputConfig';

export interface WorkflowTaskDefSchemaInput {
  id: string;
  type: string;
  config?: WorkflowTaskDefSchemaInputConfig;
  upstream?: string[];
  result_key?: string | null;
}